+++ lib/llm/active_record/acts_as_agent.rb
@@ -41,6 +41,11 @@ module LLM::ActiveRecord
         agent.concurrency(concurrency)
       end
 
+      def tracer(tracer = nil, &block)
+        return agent.tracer if tracer.nil? && !block
+        agent.tracer(tracer, &block)
+      end
+
       def agent
         @agent ||= Class.new(LLM::Agent)
       end